requirements.txt
speedtest-cli==2.0.0